+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Unreleased
+
 ## [1.1.0] - 2017-10-01
 ### Added
 - EmailAddress class.