+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## [1.2.0] - 2017-12-22
 - Added equals method to all data type classes.
 - Added toInteger method to IPAddress class.
 - Added fromInteger method to IPAddress class.
@@ -12,5 +12,6 @@
 ## 1.0.0 - 2017-05-23
 - First stable revision.
 
+[1.2.0]: https://github.com/themichaelhall/datatypes/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/themichaelhall/datatypes/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/themichaelhall/datatypes/compare/v1.0.0...v1.0.1